chore: gitignore 忽略 Reasonix 本地工具配置
.reasonix/ 桌面元数据与 reasonix.toml 权限白名单为本地开发环境
配置，不进仓库。同时移除此前误暂存的这两个路径。

+++ server/internal/model/amortization.go
@@ -8,8 +8,8 @@ type AmortizationSnapshot struct {
 	ActivityID       uint      `gorm:"uniqueIndex:idx_activity_date;not null" json:"activity_id"`
 	Date             time.Time `gorm:"uniqueIndex:idx_activity_date;not null" json:"date"` // 快照日期
 	ExecutionCount   int       `gorm:"not null" json:"execution_count"`                   // 当日执行次数
-	AmortizationBase float64   `gorm:"not null" json:"amortization_base"`                 // 有效摊销基数
-	DailyAmount      float64   `gorm:"not null" json:"daily_amount"`                      // 当日摊销金额
+	AmortizationBase int64     `gorm:"not null" json:"amortization_base"` // 有效摊销基数（单位：分）
+	DailyAmount      int64     `gorm:"not null" json:"daily_amount"`      // 当日摊销金额（单位：分）
 	CreatedAt        time.Time `json:"created_at"`
 }
